requirements/base.txt
asgiref==3.2.10
Django==3.0.7
et-xmlfile==1.0.1
html5lib==1.1
jdcal==1.4.1
openpyxl==3.0.4
Pillow==7.2.0
psycopg2-binary==2.8.5
PyPDF2==1.26.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
uuid==1.30
webencodings==0.5.1
xhtml2pdf==0.2.4
django-environ==0.4.5
python-barcode==0.9.0
